Add files via upload
This code is a classic example of a simple neural network learning the linear relationship between x and y. In this case, the relationship is y = 2x - 1. During training, the model adjusts its single weight and bias to try to learn this relationship. After training with the provided data (xs and ys), the model should be able to make reasonably accurate predictions for new x values not included in the training data, like 10.0 in this case.

Using a neural network for this kind of problem is a bit like using a sledgehammer to crack a nut, as it's a very simple problem that would typically be solved with more basic methods. However, it's a good example to illustrate the basic concepts of machine learning and neural networks.